Cambio de nombre al paquete y arreglo de paquetes
