+++ src/windowsapplication/controller/MainWindowController.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.controller;
+package windowsapplication.controller;
 
 import javafx.scene.Parent;
 import javafx.stage.Stage;
