+++ src/windowsapplication/controller/SearchCreateGroupWindowController.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.controller;
+package windowsapplication.controller;
 
 /**
  *
  * @author Gaizka Andres
  */
-public class InfoDocWindowController {
+public class SearchCreateGroupWindowController {
     
 }
