+++ src/windowsapplication/controller/UploadDocWindowController.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package windowsappication.controller;
+package windowsapplication.controller;
 
 /**
  *
  * @author Gaizka Andres
  */
-public class InfoGroupWindowController {
+public class UploadDocWindowController {
     
 }
