+++ src/windowsapplication/view/MetodoDePago.fxml
