+++ src/windowsapplication/view/VerDocumento.fxml
